stonk_bot/requirements.txt
line-bot-sdk
flask
mysql.connector